requirements.txt
faker
python-dotenv
mysql-connector-python